+++ composeApp/src/commonMain/kotlin/org/darchacheron/gofirst/settings/SettingsView.kt
@@ -12,8 +12,10 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
@@ -40,8 +42,11 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
+import com.github.skydoves.colorpicker.compose.AlphaTile
+import com.github.skydoves.colorpicker.compose.BrightnessSlider
 import com.github.skydoves.colorpicker.compose.ColorEnvelope
 import com.github.skydoves.colorpicker.compose.HsvColorPicker
 import com.github.skydoves.colorpicker.compose.rememberColorPickerController
@@ -49,6 +54,9 @@ import gofirst.composeapp.generated.resources.Res
 import gofirst.composeapp.generated.resources.ic_back
 import gofirst.composeapp.generated.resources.ic_reset
 import gofirst.composeapp.generated.resources.ic_save
+import gofirst.composeapp.generated.resources.settings_color_cancel
+import gofirst.composeapp.generated.resources.settings_color_select
+import gofirst.composeapp.generated.resources.settings_color_title
 import gofirst.composeapp.generated.resources.settings_content_description_back
 import gofirst.composeapp.generated.resources.settings_content_description_reset
 import gofirst.composeapp.generated.resources.settings_content_description_save
@@ -153,6 +161,7 @@ fun SettingsView(
 private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Settings) {
     var isColorPickerVisible by remember { mutableStateOf(false) }
     var playerId by remember { mutableStateOf<Int?>(null) }
+    var playerColor by remember { mutableStateOf<Color?>(null) }
 
     Text(
         text = stringResource(Res.string.settings_theme),
@@ -182,6 +191,7 @@ private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Sett
             .clickable(onClick = {
                 isColorPickerVisible = true
                 playerId = 0
+                playerColor = Color(settings.player1Color.toULong())
             })
         )
         Spacer(modifier = Modifier.width(16.dp))
@@ -200,6 +210,7 @@ private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Sett
             .clickable(onClick = {
                 isColorPickerVisible = true
                 playerId = 1
+                playerColor = Color(settings.player2Color.toULong())
             })
         )
         Spacer(modifier = Modifier.width(16.dp))
@@ -218,6 +229,7 @@ private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Sett
             .clickable(onClick = {
                 isColorPickerVisible = true
                 playerId = 2
+                playerColor = Color(settings.player3Color.toULong())
             })
         )
         Spacer(modifier = Modifier.width(16.dp))
@@ -236,6 +248,7 @@ private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Sett
             .clickable(onClick = {
                 isColorPickerVisible = true
                 playerId = 3
+                playerColor = Color(settings.player4Color.toULong())
             })
         )
         Spacer(modifier = Modifier.width(16.dp))
@@ -254,6 +267,7 @@ private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Sett
             .clickable(onClick = {
                 isColorPickerVisible = true
                 playerId = 4
+                playerColor = Color(settings.player5Color.toULong())
             })
         )
         Spacer(modifier = Modifier.width(16.dp))
@@ -272,6 +286,7 @@ private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Sett
             .clickable(onClick = {
                 isColorPickerVisible = true
                 playerId = 5
+                playerColor = Color(settings.player6Color.toULong())
             })
         )
         Spacer(modifier = Modifier.width(16.dp))
@@ -290,6 +305,7 @@ private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Sett
             .clickable(onClick = {
                 isColorPickerVisible = true
                 playerId = 6
+                playerColor = Color(settings.player7Color.toULong())
             })
         )
         Spacer(modifier = Modifier.width(16.dp))
@@ -308,6 +324,7 @@ private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Sett
             .clickable(onClick = {
                 isColorPickerVisible = true
                 playerId = 7
+                playerColor = Color(settings.player8Color.toULong())
             })
         )
         Spacer(modifier = Modifier.width(16.dp))
@@ -326,6 +343,7 @@ private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Sett
             .clickable(onClick = {
                 isColorPickerVisible = true
                 playerId = 8
+                playerColor = Color(settings.player9Color.toULong())
             })
         )
         Spacer(modifier = Modifier.width(16.dp))
@@ -344,6 +362,7 @@ private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Sett
             .clickable(onClick = {
                 isColorPickerVisible = true
                 playerId = 9
+                playerColor = Color(settings.player10Color.toULong())
             })
         )
         Spacer(modifier = Modifier.width(16.dp))
@@ -356,41 +375,67 @@ private fun SettingsControl(settingsViewModel: SettingsViewModel, settings: Sett
     if (isColorPickerVisible) {
         ColorPickerDialog(
             playerId = playerId!!,
+            initialColor = playerColor!!,
             onDismiss = {
                 isColorPickerVisible = false
                 playerId = null
+                playerColor = null
             },
             onSave = { playerId, color -> settingsViewModel.onPlayerColorChanged(playerId, color.value.toLong()) })
     }
 }
 
 @Composable
-private fun ColorPickerDialog(playerId: Int, onDismiss:() -> Unit, onSave: (playerId: Int, color: Color) -> Unit) {
+private fun ColorPickerDialog(playerId: Int, initialColor: Color, onDismiss:() -> Unit, onSave: (playerId: Int, color: Color) -> Unit) {
     val controller = rememberColorPickerController()
+    val scrollState = rememberScrollState()
 
     AlertDialog(
         onDismissRequest = { onDismiss() },
-        title = { Text(text = "Pick a color") },
+        title = { Text(text = stringResource(Res.string.settings_color_title)) },
         text = {
-            HsvColorPicker(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(450.dp)
-                    .padding(10.dp),
-                controller = controller
-            )
+            Column(modifier = Modifier.fillMaxSize().verticalScroll(scrollState)) {
+                Box(modifier = Modifier.weight(5f).fillMaxWidth()) {
+                    HsvColorPicker(
+                        modifier = Modifier
+                            .align(Alignment.TopCenter)
+                            .fillMaxWidth()
+                            .padding(10.dp),
+                        controller = controller,
+                        initialColor = initialColor
+                    )
+                }
+
+                BrightnessSlider(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(10.dp)
+                        .height(35.dp)
+                        .align(Alignment.CenterHorizontally),
+                    controller = controller,
+                )
+
+                AlphaTile(
+                    modifier = Modifier
+                        .size(80.dp)
+                        .clip(RoundedCornerShape(6.dp))
+                        .align(Alignment.CenterHorizontally),
+                    controller = controller,
+                )
+                Spacer(modifier = Modifier.weight(1f))
+            }
         },
         confirmButton = {
             Button(onClick = {
                 onSave(playerId, controller.selectedColor.value)
                 onDismiss()
             }) {
-                Text(text = "Select")
+                Text(text = stringResource(Res.string.settings_color_select))
             }
         },
         dismissButton = {
             TextButton(onClick = onDismiss) {
-                Text(text = "Cancel")
+                Text(text = stringResource(Res.string.settings_color_cancel))
             }
         }
     )